pyproject.toml
[tool.poetry]
name = "document-classifier"
version = "0.1.0"
description = "Document classification service using FastAPI and OpenRouter AI"
authors = ["foundanand"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
PyPDF2 = "^3.0.1"
openai = "^1.3.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"